+++ docs/conf.py
@@ -361,6 +361,9 @@ needs_render_context = {
 # build needs.json to make permalinks work
 needs_build_json = True
 
+# build needs_json for every needs-id to make detail panel
+needs_build_json_per_id = False
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]